fix: elimination of remarks of the mentor
