+++ react-app/src/App.tsx
@@ -1,20 +1,23 @@
-import AboutPage from './pages/AboutPage';
-import HomePage from './pages/HomePage';
-import NotFoundPage from './pages/NotFoundPage';
+import React, { Component } from 'react';
+import { AboutPage } from './pages/AboutPage';
+import { HomePage } from './pages/HomePage';
+import { NotFoundPage } from './pages/NotFoundPage';
 import { Route, Routes } from 'react-router-dom';
-import Header from './components/Header';
+import { Header } from './components/Header';
 
-function App() {
-  return (
-    <>
-      <Header />
-      <Routes>
-        <Route path={'/'} element={<HomePage />} />
-        <Route path={'/about'} element={<AboutPage />} />
-        <Route path={'*'} element={<NotFoundPage />} />
-      </Routes>
-    </>
-  );
+class App extends Component {
+  render() {
+    return (
+      <>
+        <Header />
+        <Routes>
+          <Route path={'/'} element={<HomePage />} />
+          <Route path={'/about'} element={<AboutPage />} />
+          <Route path={'*'} element={<NotFoundPage />} />
+        </Routes>
+      </>
+    );
+  }
 }
 
-export default App;
+export { App };
