+++ react-app/src/pages/AboutPage.tsx
@@ -10,4 +10,4 @@ class AboutPage extends React.Component {
   }
 }
 
-export default AboutPage;
+export { AboutPage };
