+++ react-app/src/pages/NotFoundPage.tsx
@@ -17,4 +17,4 @@ class NotFoundPage extends React.Component {
   }
 }
 
-export default NotFoundPage;
+export { NotFoundPage };
