+++ react-app/src/tests/NotFoundPage.test.tsx
@@ -1,0 +1,14 @@
+import { render, screen } from '@testing-library/react';
+import { NotFoundPage } from 'pages/NotFoundPage';
+import { BrowserRouter } from 'react-router-dom';
+describe('NotFoundPage', () => {
+  it('renders correctly', () => {
+    render(
+      <BrowserRouter>
+        <NotFoundPage />
+      </BrowserRouter>
+    );
+    const notFoundElement = screen.getByText(/PAGE NOT FOUND/i);
+    expect(notFoundElement).toBeInTheDocument();
+  });
+});
