+++ react-app/vite.config.ts
@@ -4,7 +4,6 @@ import react from '@vitejs/plugin-react';
 import { defineConfig } from 'vite';
 import eslintPlugin from 'vite-plugin-eslint';
 import tsconfigPaths from 'vite-tsconfig-paths';
-// https://vitejs.dev/config/
 
 export default defineConfig({
   plugins: [eslintPlugin(), react(), tsconfigPaths()],
